+++ packages/fighting-design/close-btn/src/close-btn.ts
@@ -1,4 +1,4 @@
-import type { ExtractPropTypes, PropType } from 'vue'
+import type { ExtractPropTypes, PropType, VNode } from 'vue'
 
 export const Props = {
   size: {
@@ -16,6 +16,10 @@ export const Props = {
   color: {
     type: String,
     default: (): string => ''
+  },
+  icon: {
+    type: Object as PropType<VNode>,
+    default: (): null => null
   }
 } as const