+++ start/src/App.vue
@@ -7,9 +7,9 @@
 </script>
 
 <template>
-  <f-button type="default" loading>
-    默认按钮
-  </f-button>
+  <f-alert type="primary" close title="哈哈哈哈哈">主要提示信息</f-alert>
+  <f-alert type="success" close>成功提示信息</f-alert>
+  <f-alert type="danger" close>危险提示信息</f-alert>
 </template>
 
 <style lang="scss" scoped></style>