+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 - 新增 `f-input` 组件
 - 新增 `f-swap` 组件
+- `f-button` 组件 `before-icon` `after-icon` `loading-icon` 配置项改为 `VNode` 类型，默认值修改为 `null`
+- `f-close-btn` 新增 `icon` 配置项
+
 - 废除 `f-dialog` 组件 `close-icon` 属性
 - 新增 `f-dialog` 组件 `closeIcon` 插槽
 - 废除 `f-drawer` 组件 `close-icon` 属性
@@ -11,7 +14,6 @@
 - 废除 `f-alert` 组件 `icon` 属性
 - 新增 `f-alert` 组件 `prefixIcon` 和 `closeIcon` 插槽
 - 新增 `f-close-btn` 组件 `color` 属性
-- `f-button` 组件 `before-icon` `after-icon` `loading-icon` 配置项改为 `VNode` 类型，默认值修改为 `null`
 - 废除 `f-ink` 组件 `before-icon` `after-icon` 配置项
 - 新增 `f-ink` 组件 `beforeIcon` `afterIcon` 插槽
 - 废除 `f-avatar` 组件 `icon` 配置项